+++ Package@swift-6.1.swift
@@ -1,0 +1,128 @@
+// swift-tools-version:6.1
+#if canImport(Darwin)   
+import Darwin.C
+#elseif canImport(Glibc)
+import Glibc
+#elseif canImport(MSVCRT)
+import MSVCRT
+#endif
+
+import PackageDescription
+
+var products: [Product] = [
+    .library(name: "Sentry", targets: ["Sentry", "SentryCppHelper"]),
+    .library(name: "Sentry-Dynamic", targets: ["Sentry-Dynamic"]),
+    .library(name: "Sentry-Dynamic-WithARM64e", targets: ["Sentry-Dynamic-WithARM64e"]),
+    .library(name: "Sentry-WithoutUIKitOrAppKit", targets: ["Sentry-WithoutUIKitOrAppKit", "SentryCppHelper"]),
+    .library(name: "Sentry-WithoutUIKitOrAppKit-WithARM64e", targets: ["Sentry-WithoutUIKitOrAppKit-WithARM64e", "SentryCppHelper"]),
+    .library(name: "SentrySwiftUI", targets: ["Sentry", "SentrySwiftUI", "SentryCppHelper"])
+]
+
+var targets: [Target] = [
+    .binaryTarget(
+        name: "Sentry",
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/8.58.0/Sentry.xcframework.zip",
+        checksum: "d883f19575ab064633237c2c58ee16a1e596c58f5efe1c110f65998bc71b0389" //Sentry-Static
+    ),
+    .binaryTarget(
+        name: "Sentry-Dynamic",
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/8.58.0/Sentry-Dynamic.xcframework.zip",
+        checksum: "07d2a8a928bf49d44ed498a4a994d736d3e9e44667e5c5082dba0d6be526b339" //Sentry-Dynamic
+    ),
+    .binaryTarget(
+        name: "Sentry-Dynamic-WithARM64e",
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/8.58.0/Sentry-Dynamic-WithARM64e.xcframework.zip",
+        checksum: "2f13dc949fb3f795222f7420aca649b32f365137d27cc13311e2e59eeb902691" //Sentry-Dynamic-WithARM64e
+    ),
+    .binaryTarget(
+        name: "Sentry-WithoutUIKitOrAppKit",
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/8.58.0/Sentry-WithoutUIKitOrAppKit.xcframework.zip",
+        checksum: "d6e65593d9538db22d318afda5828889d3eacd1aa89957a37fe81725ec91141e" //Sentry-WithoutUIKitOrAppKit
+    ),
+    .binaryTarget(
+        name: "Sentry-WithoutUIKitOrAppKit-WithARM64e",
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/8.58.0/Sentry-WithoutUIKitOrAppKit-WithARM64e.xcframework.zip",
+        checksum: "93377c7cb7a020c5bab375ca32bd26f954e54dcb1b21860b9c6dff0f3476413e" //Sentry-WithoutUIKitOrAppKit-WithARM64e
+    ),
+    .target (
+        name: "SentrySwiftUI",
+        dependencies: ["Sentry", "SentryInternal"],
+        path: "Sources/SentrySwiftUI",
+        exclude: ["SentryInternal/", "module.modulemap"],
+        linkerSettings: [
+            .linkedFramework("Sentry")
+        ]),
+    .target(
+        name: "SentryInternal",
+        path: "Sources/SentrySwiftUI",
+        sources: [
+            "SentryInternal/"
+        ],
+        publicHeadersPath: "SentryInternal/"),
+    .target(
+        name: "SentryCppHelper",
+        path: "Sources/SentryCppHelper",
+        linkerSettings: [
+         .linkedLibrary("c++")
+        ]
+    )
+]
+
+let env = getenv("EXPERIMENTAL_SPM_BUILDS")
+if let env = env, String(cString: env, encoding: .utf8) == "1" {
+    products.append(.library(name: "SentrySPM", type: .dynamic, targets: ["SentryObjc"]))
+    targets.append(contentsOf: [
+        // At least one source file is required, therefore we use a dummy class to satisfy the SPM build system
+        .target(
+            name: "SentryHeaders",
+            path: "Sources/Sentry", 
+            sources: ["SentryDummyPublicEmptyClass.m"],
+            publicHeadersPath: "Public"
+        ),
+        .target(
+            name: "_SentryPrivate",
+            dependencies: ["SentryHeaders"],
+            path: "Sources/Sentry",
+            sources: ["SentryDummyPrivateEmptyClass.m"],
+            publicHeadersPath: "include",
+            cSettings: [.headerSearchPath("include/HybridPublic")]),
+        .target(
+            name: "SentrySwift",
+            dependencies: ["_SentryPrivate", "SentryHeaders"],
+            path: "Sources/Swift",
+            swiftSettings: [
+                .unsafeFlags(["-enable-library-evolution"]),
+                // Some API breaking changes are necessary for the framework to compile with SPM, we’ll ship
+                // those in V9.
+                .define("SDK_V9")
+            ]),
+        .target(
+            name: "SentryObjc",
+            dependencies: ["SentrySwift"],
+            path: "Sources",
+            exclude: ["Sentry/SentryDummyPublicEmptyClass.m", "Sentry/SentryDummyPrivateEmptyClass.m", "Swift", "SentrySwiftUI", "Resources", "Configuration", "SentryCppHelper"],
+            publicHeadersPath: "Sentry/include",
+            cSettings: [
+                .headerSearchPath("Sentry/include/HybridPublic"),
+                .headerSearchPath("Sentry"),
+                .headerSearchPath("SentryCrash/Recording"),
+                .headerSearchPath("SentryCrash/Recording/Monitors"),
+                .headerSearchPath("SentryCrash/Recording/Tools"),
+                .headerSearchPath("SentryCrash/Installations"),
+                .headerSearchPath("SentryCrash/Reporting/Filters"),
+                .headerSearchPath("SentryCrash/Reporting/Filters/Tools"),
+                .define("SDK_V9")])
+    ])
+}
+
+let package = Package(
+    name: "Sentry",
+    platforms: [.iOS(.v11), .macOS(.v10_13), .tvOS(.v11), .watchOS(.v4)],
+    products: products,
+    traits: [
+        .default(enabledTraits: [])
+    ],
+    targets: targets,
+    swiftLanguageModes: [.v5],
+    cxxLanguageStandard: .cxx14
+)